+++ golem/managers/demand/refreshing.py
@@ -10,8 +10,7 @@ from golem.payload import Payload
 from golem.payload import defaults as payload_defaults
 from golem.resources import Allocation, Demand, Proposal
 from golem.resources.demand.demand_builder import DemandBuilder
-from golem.utils.asyncio import create_task_with_logging
-from golem.utils.asyncio.queue import ErrorReportingQueue
+from golem.utils.asyncio import ErrorReportingQueue, create_task_with_logging
 from golem.utils.logging import get_trace_id_name, trace_span
 
 logger = logging.getLogger(__name__)
